+++ Source/Components/GAMs/SimulinkWrapperGAM/SimulinkWrapperGAM.h
@@ -610,7 +610,7 @@ private:
      * @param[in] depth       tree depth level at call (for recursive callings)
      * @returns `true` if the parameter has been correctly scanned, `false` otherwise.
      */
-    bool ScanParameter(const uint32 parIdx, StreamString spacer, const ParameterMode mode, void* const startAddress, StreamString baseName, const uint64 baseOffset, const uint32 depth);
+    bool ScanParameter(const uint16 parIdx, StreamString spacer, const ParameterMode mode, void* const startAddress, StreamString baseName, const uint64 baseOffset, const uint32 depth);
 
     /**
      * @brief Scans the root level input or output tree of the loaded Simulink .so code
